Update SampleApp.py
Implemented "Forgot Password" functionality. User now has the ability to enter their username, and the system will query the database using that username, and retrieve an email address to send a temporary password to. At the same time that password will be updated on the "employees" table so that the user is able to log in with that temporary password. 

Also, "Login" button now authenticates AND switches to the "HomePage" widget screen. Before it was only authenticating, but not advancing to the new screen.